Stripe API 2019-12-03 Remove deprecated Customer tax_info
Deprecated tax information for Customers have been removed.
The deprecated tax_info and tax_info_verification fields on the Customer object are now removed in favor of tax_ids.
The deprecated tax_info parameter on the Customer create and update methods are removed in favor of tax_id_data.
For more information, view the migration guide.